test: Criado teste de validacao de calculo de preco medio ponderado de compra

+++ system/domain/value_objects/returns_value_objct.py
@@ -6,7 +6,7 @@ from system.domain.value_objects.base_value_object import BaseValueObject
 
 class ReturnsValueObject(BaseValueObject):
     average_price: Decimal
-    quantity: Decimal
+    quantity: int
     total_value: Decimal
     loss: Optional[Decimal]
     returns: Decimal
